ft: added delete error handling to search component

+++ src/views/Product.js
@@ -1,7 +1,8 @@
 import React, {useState, useEffect} from "react";
 import Header from "./Header";
 import {Table} from "react-bootstrap"
-import {Link, useParams} from "react-router-dom";
+import {Link, useParams, useNavigate} from "react-router-dom";
+
 
 
 function ProductList() {
@@ -9,6 +10,7 @@ function ProductList() {
     const [data,setData] = useState([]);
     const {id} = useParams();
     const [errors,setErrors] = useState(null)
+    const navigate = useNavigate();
 
     useEffect(() => {
         async function fetchData() {
@@ -34,10 +36,22 @@ function ProductList() {
     },[])
 
     async function deleteOperation(id) {
-        let result = await fetch("http://localhost:8000/api/delete/" + id, {
-            method: "DELETE"
-        });
-        result = await result.json();
+        if(window.confirm("Do you really want to delete this product?")) {
+            try {
+                let result = await fetch("http://localhost:8000/api/delete/" + id, {
+                    method: "DELETE"
+                });
+                if(!result.ok) {
+                    const dataError = await result.json();
+                    throw new Error(dataError.message);
+                }
+                alert("The product has been deleted");
+                navigate("/")
+            
+            }catch(error) {
+                alert("Error " + error.message);
+            }
+        }
         
     }
 
@@ -60,7 +74,7 @@ function ProductList() {
         <>
             <Header />
             <h1 className="title">{data.name}</h1>
-            <div className="col-sm-8 offset-sm-2 product-card">
+            <div className="product-card">
                 <div id="container-img-product">
                     <img src={"http://localhost:8000/storage/" + data.file_path} alt={"Photography of " + data.description}></img>
                 </div>
@@ -70,7 +84,7 @@ function ProductList() {
                     <p>Characteristics of the product: {data.description}</p>
                     <div className="product-btns">
                         <button className="btn btn-danger" onClick={() => deleteOperation(data.id)}>Delete</button>
-                        <Link className="btn btn-secondary"  to={"updateproduct/" + data.id}>Edit</Link>
+                        <Link className="btn btn-secondary"  to={"/updateproduct/" + data.id}>Edit</Link>
                     </div>
                                
                 </div>
